src/gradio_app/requirements.txt
python-dotenv==1.0.0
gradio